+++ CHANGELOG.md
@@ -25,6 +25,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/), and t
 
 ### 👷 Changed
 
+- Change Dialog (feedback) core component
+  - Add async support to action and cancel button
 - Change Theme (layout) core component
   - Add smoothscroll polyfill
 - Change Checkbox (input) core component